fix: data 파일 분리
