+++ src/app/components/LoginForm/LoginForm.tsx
@@ -1,45 +1,8 @@
 import React, { useState } from 'react';
-import { Button, Input, FlexBox, InputStatus } from '@elements';
-import IconCheckGreen from '@assets/images/icons/icon-check-green.svg';
-import IconPasswordShow from '@assets/images/icons/icon-password-show.svg';
-import IconPasswordHide from '@assets/images/icons/icon-password-hide.svg';
+import { Button, Input, FlexBox } from '@elements';
+import { checkValidation } from '@libs/util';
 import palette from '@libs/theme/palettes';
-
-interface ILoginInfo {
-  name: string;
-  type: 'email' | 'password' | 'text';
-  status: InputStatus;
-  isValidated: boolean;
-  value: string;
-  placeholder: string;
-  icon: string[];
-  errorMessage: string;
-  activeMessage?: string;
-  inActiveMessage?: string;
-}
-
-const LOGIN_INFO: ILoginInfo[] = [
-  {
-    name: 'email',
-    type: 'email',
-    status: 'inActive',
-    isValidated: false,
-    value: '',
-    placeholder: '이메일',
-    icon: [IconCheckGreen],
-    errorMessage: '올바른 이메일 형식을 입력해주세요.',
-  },
-  {
-    name: 'password',
-    type: 'password',
-    status: 'inActive',
-    isValidated: false,
-    value: '',
-    placeholder: '비밀번호',
-    icon: [IconPasswordHide, IconPasswordShow],
-    errorMessage: '비밀번호를 다시 확인해주세요.',
-  },
-];
+import { LOGIN_INFO, type ILoginInfo } from './loginForm.data';
 
 function LoginForm() {
   // prop destruction
@@ -53,15 +16,6 @@ function LoginForm() {
   // calculated values
   // effects
   // handlers
-
-  const checkValidation = (type: string, value: string) => {
-    const pattern: Record<string, RegExp> = {
-      email: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
-      password: /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,16}$/,
-    };
-
-    return pattern[type].test(value);
-  };
 
   const checkInputInfo = (event: React.ChangeEvent<HTMLInputElement>, inputName: string) => {
     const targetValue = event.currentTarget.value;
