+++ src/app/libs/util/checkValidation.ts
@@ -1,0 +1,9 @@
+export const checkValidation = (type: string, value: string) => {
+  const pattern: Record<string, RegExp> = {
+    email: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
+    password: /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,16}$/,
+    confirmPassword: /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,16}$/,
+  };
+
+  return pattern[type].test(value);
+};
