+++ src/app/libs/util/index.ts
@@ -1,1 +1,2 @@
+export * from './checkValidation';
 export * from './getToken';
